Inform user about CorruptDatabaseBackup.db
Prior to this commit, the user gets no feedback what-so-ever of what is
going on when a corrupted database gets renamed. Which we should assume
can be a quite horrifying experience.

This adds a friendly message about what has happened, and suggests what
to do.

GitHub issue #6212, named "More useful error message for corrupted
databases", might still be relevant for the "more useful" part.

By reusing the design and making the dual-alternative of HomeFragment
into instead one of three container views, the code changes stay small
and complexity low. Easy to merge, and easy to revert once some code for
a better implementation actually gets available.

The date parsing, used when logging detection of a corrupt database, is
copied verbatim from CrashReportWriter.java which presumably should make
it reasonable enough.